refactor: 고객 쿠폰 조회 요청 분리
